+++ tests/testthat/test-cleaning-utils.R
@@ -97,3 +97,12 @@ test_that("Fix make state abbreviations works", {
 
   expect_true(all(is.na(make_state_abb(1:5))))
 })
+
+test_that("Get Mode works properly", {
+  expect_true(is.na(get_mode(1:5)))
+  expect_equal(get_mode(c(1, 1, 2, 3, 4)), 1)
+  expect_equal(get_mode(c(1, 1:5, NA, NA, NA)), 1)
+  expect_true(is.na(get_mode(c(1, 1:5, NA, NA, NA), remove_NA = FALSE)))
+  expect_equal(get_mode(c(1, 1, 2, 2, 4, 5)), c(1, 2))
+  expect_equal(get_mode(c(1, 1, 3, 3, 2, 2, 5, 6)), c(1, 2, 3))
+})